requirements.txt
flask>=2.3.0
requests>=2.31.0
PyJWT>=2.8.0
